requirements.txt
# AfterDark Validation Kit - Python Dependencies

# HTTP/API Testing
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Cloud Provider SDKs
oci>=2.117.0
boto3>=1.33.0
cloudflare>=2.11.0

# Database Testing
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Security Testing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.0
rich>=13.7.0
tabulate>=0.9.0

# Async Support
asyncio>=3.4.3

# JSON/Data Processing
jsonschema>=4.20.0
pydantic>=2.5.0

# Reporting
jinja2>=3.1.0
markdown>=3.5.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
